feat: add css for blogs
