+++ posts/functions-for-beginners.mdx
@@ -1,19 +1,22 @@
 ---
-title: Functions for Beginners
-date: 'August 4, 2021'
-description: Learn how to create and use functions
-thumbnailUrl: '/javascript-functions-thumbnail.jpeg'
-tags: ['functions', 'javascript']
+title: 'How to set up Tailwind CSS'
+date: 'September 2, 2021'
+description: "Here's a short guide on setting up Tailwind CSS"
+thumbnailUrl: '/tailwindcss-thumbnail.jpg'
+tags: ['tailwindcss', 'styling']
+author: 'Chirag Kale'
+readingTime: '3 min'
 ---
 
 <div>
-  This is a blog post about <strong>functions</strong>. <br/>
-  We'll learn what functions are, how to write functions, and how to use them.
+  This is a blog post about <strong>Tailwind CSS</strong>. <br />
+  We'll learn how to set up and use Tailwind CSS.
 </div>
 
-<br/>
+<br />
 
 ### This is a Markdown sub-heading
+
 We just used a `###` subheading in our blog post.
 
 <SyntaxHighlighter language="javascript">
@@ -30,5 +33,3 @@ We just used a `###` subheading in our blog post.
     return console.log('hey')
   }
 ```
-
-<div>Here, I'm using a component inside my MarkDown file: {<Button text={"Click me"}/>}</div>
