+++ posts/solidjs-for-beginners.mdx
@@ -1,18 +1,22 @@
 ---
-title: 'SolidJS: A Beginners Guide'
-date: 'September 11, 2021'
-description: "Here's an introduction to SolidJS"
-thumbnailUrl: "/solidjs-thumbnail.jpeg"
-tags: ['JavaScript', 'solidjs']
+title: 'How to set up Tailwind CSS'
+date: 'September 2, 2021'
+description: "Here's a short guide on setting up Tailwind CSS"
+thumbnailUrl: '/tailwindcss-thumbnail.jpg'
+tags: ['tailwindcss', 'styling']
+author: 'Chirag Kale'
+readingTime: '3 min'
 ---
+
 <div>
-  This is a blog post about <strong>SolidJS</strong>. <br/>
-  We'll learn about SolidJS and how to build with it.
+  This is a blog post about <strong>Tailwind CSS</strong>. <br />
+  We'll learn how to set up and use Tailwind CSS.
 </div>
 
-<br/>
+<br />
 
 ### This is a Markdown sub-heading
+
 We just used a `###` subheading in our blog post.
 
 <SyntaxHighlighter language="javascript">
@@ -29,5 +33,3 @@ We just used a `###` subheading in our blog post.
     return console.log('hey')
   }
 ```
-
-<div>Here, I'm using a component inside my MarkDown file: {<Button text={"Click me"}/>}</div>
